Move the LHN performance tests to the existing Sidebar suites
